requirements.txt
nptdms
dvc
dvc_gdrive
pandas

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
